Move managed-auth args to config package
